datastreaming/requirements.txt
kafka-python
pymongo
pandas
neo4j